# src/CashFlow.Api/CashFlowApi.Docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CashFlow.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5162/"
    }
  ],
  "paths": {
...
    "/api/Reports/excel": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "month",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/Reports/pdf": {
      "get": {
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "month",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
...
}